Move CKEditor5 to containers because it has a registry now
