+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/CKEditor5/CKEditor5.js
