+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/CKEditor5/index.js
