+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/CKEditor5/plugins/InternalLinkPlugin/registries/InternalLinkTypeRegistry.js
