+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/CKEditor5/plugins/InternalLinkPlugin/tests/registries/InternalLinkTypeRegistry.test.js
